feature: add Explicit attribute to all test cases
